Move DeviceBrowser to components
